Reduced logging by BlockManager in default log4j.properties
also fixed a putInto hydra test and update spark module link

+++ dtests/src/resources/scripts/putInto/create_and_load_testLTable.sql
@@ -1,4 +1,4 @@
 -- DROP TABLE IF ALREADY EXISTS --
 DROP TABLE IF EXISTS testL;
 
-create table testL (id long, data string, data2 decimal) using column options (partition_by 'id', key_columns 'id');
+create table testL (id long, data string, data2 decimal(38,10)) using column options (partition_by 'id', key_columns 'id');
